requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
langchain==0.2.0
langchain-google-genai
google-generativeai
openai-whisper
ffmpeg
yt-dlp
celery[redis]==5.3.0
redis==4.5.0